modifiche

+++ resources/views/users/index.blade.php
@@ -27,23 +27,45 @@
             </thead>
             <tbody>
             @foreach($users as $user)
-                <tr>
-                    <th scope="row">{{ $user->id }}</th>
-                    <td>{{ $user->name }}</td>
-                    <td>{{ $user->email }}</td>
-                    <td>
-                        @foreach($user->roles as $role)
-                            <span class="badge bg-primary">{{ $role->name }}</span>
-                        @endforeach
-                    </td>
-                    <?php // <td><a href="{{ route('users.show', $user->id) }}" class="btn btn-warning btn-sm">Vedi</a></td> ?>
-                    <td><a href="{{ route('users.edit', $user->id) }}" class="btn btn-info btn-sm">Modifica</a></td>
-                    <td>
-                        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-                        {!! Form::submit('Elimina', ['class' => 'btn btn-danger btn-sm']) !!}
-                        {!! Form::close() !!}
-                    </td>
-                </tr>
+                @if($isSuperAdmin == 1)
+                    <tr>
+                        <th scope="row">{{ $user->id }}</th>
+                        <td>{{ $user->name }}</td>
+                        <td>{{ $user->email }}</td>
+                        <td>
+                            @foreach($user->roles as $role)
+                                <span class="badge bg-primary">{{ $role->name }}</span>
+                            @endforeach
+                        </td>
+                        <?php // <td><a href="{{ route('users.show', $user->id) }}" class="btn btn-warning btn-sm">Vedi</a></td> ?>
+                        <td><a href="{{ route('users.edit', $user->id) }}" class="btn btn-info btn-sm">Modifica</a></td>
+                        <td>
+                            {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                            {!! Form::submit('Elimina', ['class' => 'btn btn-danger btn-sm']) !!}
+                            {!! Form::close() !!}
+                        </td>
+                    </tr>
+                @else
+                    @if($user->roles[0]->id !== 1)
+                        <tr>
+                            <th scope="row">{{ $user->id }}</th>
+                            <td>{{ $user->name }}</td>
+                            <td>{{ $user->email }}</td>
+                            <td>
+                                @foreach($user->roles as $role)
+                                    <span class="badge bg-primary">{{ $role->name }}</span>
+                                @endforeach
+                            </td>
+                            <?php // <td><a href="{{ route('users.show', $user->id) }}" class="btn btn-warning btn-sm">Vedi</a></td> ?>
+                            <td><a href="{{ route('users.edit', $user->id) }}" class="btn btn-info btn-sm">Modifica</a></td>
+                            <td>
+                                {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                                {!! Form::submit('Elimina', ['class' => 'btn btn-danger btn-sm']) !!}
+                                {!! Form::close() !!}
+                            </td>
+                        </tr>
+                    @endif
+                @endif
             @endforeach
             </tbody>
         </table>
